Gameshark refactor (#1867)
* first round of GameShark.cs cleanups, redundant code and stylecop cleanup

* simplify how gba readonly values are set

* a few more cleanups

* GameShark - use NESGameGenieDecoder for nes logic

* remove now unused NESGameGeneie tool

* Client.Common - move nes encode/decoder to a cheats folder

* GameShark - move genesis game genie coding logic to its own class

* remove NesGameGenieEncoder that i snow unused

* fix naming

* remove GenGameGenie in favor of using the unified Cheat Code decoder

* move Gameboy/GameGear decoding logic from GameShark to its own class, cleanup

* remove GBGameGenie, was only exposed via the toolbox, GameShark tool does all of this logic

* GameShark - move snes game genie decoding logic to its own class

* cleanup

* remove unused snes GameGenie tool

* move Gb GameShark decoding to its own class

* cleanup

* move Genesis action replay decoding logic to its own class

* cleanup

* cleanup

* cleanup

* cleanup

* move Sms ActionReplay decoding to its own class, cleanlup

* cleanup

* move snes action replay decoding to its own class

* cleanup with a helper method

* cleanup

* more cleanup

* cleanup

* cleanup

* cleanup

* cleanup

* cleanup

* move Psx GameShark decoding to its own class

* Move saturn gameshark parsing to its own class

* cleanup

* cleanup

* move N64 decoder logic to its own class

* remove GBA cheat code that hasn't been ready for 5 years, easier to try to implement this in the re-architected code, so delete for now at least

* start to detangle GBA gameshark mess

* Gameshark - remove most gba cheat code functionality for now, easier to rebuild it later

* Gameshark - cleanup

* cleanup

* cleanup

* cleanup

* redesign nes decoder to simpler design

* simplify more

* convert psx and saturn decoders and simplify

* convert more decoders

* convert more decoders

* convert more decoders

* convert more decoders

* convert gba decoder

* convert more decoders

* convert and simplify N64 decoder

* GameShark - simplify more

+++ BizHawk.Client.Common/lua/EmuLuaLibrary.NES.cs
@@ -4,6 +4,7 @@ using System.Linq;
 
 using NLua;
 
+using BizHawk.Client.Common.cheats;
 using BizHawk.Emulation.Common;
 using BizHawk.Emulation.Cores.Nintendo.NES;
 using BizHawk.Emulation.Cores.Consoles.Nintendo.QuickNES;
@@ -44,19 +45,11 @@ namespace BizHawk.Client.Common
 		{
 			if (NESAvailable && MemoryDomains != null)
 			{
-				var decoder = new NESGameGenieDecoder(code);
-				var watch = Watch.GenerateWatch(
-					MemoryDomains["System Bus"],
-					decoder.Address,
-					WatchSize.Byte,
-					DisplayType.Hex,
-					false,
-					code);
-
-				Global.CheatList.Add(new Cheat(
-					watch,
-					decoder.Value,
-					decoder.Compare));
+				var result = NesGameGenieDecoder.Decode(code);
+				if (result.IsValid)
+				{
+					Global.CheatList.Add(result.ToCheat(MemoryDomains.SystemBus, ""));
+				}
 			}
 		}
 
@@ -126,9 +119,12 @@ namespace BizHawk.Client.Common
 		{
 			if (NESAvailable)
 			{
-				var decoder = new NESGameGenieDecoder(code);
-				Global.CheatList.RemoveRange(
-					Global.CheatList.Where(c => c.Address == decoder.Address));
+				var decoder = NesGameGenieDecoder.Decode(code);
+				if (decoder.IsValid)
+				{
+					Global.CheatList.RemoveRange(
+						Global.CheatList.Where(c => c.Address == decoder.Address));
+				}
 			}
 		}
